CommPlexAPI/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.0
twilio>=9.0.0
google-genai>=1.0.0
google-cloud-firestore>=2.16.0
google-cloud-texttospeech>=2.16.0
httpx>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0
google-cloud-speech
google-cloud-aiplatform>=1.38.0
